+++ src/js/components/ListTracks/ListTracks.js
@@ -16,7 +16,7 @@ import style from './ListTracks.module.scss';
 // COMPONENT
 // ======================================================================
 
-const ListTracks = ({ variant, albumId, playlistId, entries }) => {
+const ListTracks = ({ variant, albumId, playlistId, playingOrder, entries }) => {
   const dispatch = useDispatch();
 
   const playerPlaying = useSelector(({ playerModel }) => playerModel.playerPlaying);
@@ -31,8 +31,10 @@ const ListTracks = ({ variant, albumId, playlistId, entries }) => {
 
   const optionShowFullTitles = useSelector(({ sessionModel }) => sessionModel.optionShowFullTitles);
   const optionShowStarRatings = useSelector(({ sessionModel }) => sessionModel.optionShowStarRatings);
+  const sortPlaylistTracks = useSelector(({ sessionModel }) => sessionModel.sortPlaylistTracks);
 
   const trackDetail = playingTrackList?.[playingTrackKeys[playingTrackIndex]];
+  const sortKey = (variant === 'playlists' && sortPlaylistTracks[playlistId]) || null;
 
   const totalDiscs = useMemo(() => {
     if (variant === 'albums') {
@@ -42,6 +44,14 @@ const ListTracks = ({ variant, albumId, playlistId, entries }) => {
     }
     return 1;
   }, [variant, entries]);
+
+  const handleSortTracks = (event) => {
+    const sortKey = event.currentTarget.dataset.sort;
+    dispatch.sessionModel.setSortPlaylistTracks({
+      playlistId,
+      sortKey,
+    });
+  };
 
   // scroll to playing track, if required
   useEffect(() => {
@@ -82,15 +92,90 @@ const ListTracks = ({ variant, albumId, playlistId, entries }) => {
               [style.headerWithRating]: optionShowStarRatings,
             })}
           >
-            <div>
+            <div onClick={handleSortTracks} data-sort="sortOrder">
               <span className={style.minCenter}>#</span>
             </div>
-            <div>Title</div>
+            <div onClick={handleSortTracks} data-sort="title">
+              Title
+              {sortKey === 'title-asc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowDownIcon" cover stroke />
+                </span>
+              )}
+              {sortKey === 'title-desc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowUpIcon" cover stroke />
+                </span>
+              )}
+            </div>
             <div></div>
-            <div>Artist</div>
-            <div>Album</div>
-            {optionShowStarRatings && <div className={style.headerRating}>Rating</div>}
-            <div>Duration</div>
+            <div onClick={handleSortTracks} data-sort="artist">
+              Artist
+              {sortKey === 'artist-asc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowDownIcon" cover stroke />
+                </span>
+              )}
+              {sortKey === 'artist-desc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowUpIcon" cover stroke />
+                </span>
+              )}
+            </div>
+            <div onClick={handleSortTracks} data-sort="album">
+              Album
+              {sortKey?.startsWith('album-asc') && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowDownIcon" cover stroke />
+                </span>
+              )}
+              {sortKey?.startsWith('album-desc') && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowUpIcon" cover stroke />
+                </span>
+              )}
+            </div>
+            {optionShowStarRatings && (
+              <div className={style.headerRating} onClick={handleSortTracks} data-sort="userRating">
+                Rating
+                {sortKey === 'userRating-desc' && (
+                  <span className={style.sortIcon}>
+                    <Icon icon="ArrowUpIcon" cover stroke />
+                  </span>
+                )}
+                {sortKey === 'userRating-asc' && (
+                  <span className={style.sortIcon}>
+                    <Icon icon="ArrowDownIcon" cover stroke />
+                  </span>
+                )}
+              </div>
+            )}
+            {/* <div onClick={handleSortTracks} data-sort="addedAt">
+              Added
+              {sortKey === 'addedAt-asc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowDownIcon" cover stroke />
+                </span>
+              )}
+              {sortKey === 'addedAt-desc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowUpIcon" cover stroke />
+                </span>
+              )}
+            </div> */}
+            <div onClick={handleSortTracks} data-sort="duration">
+              Duration
+              {sortKey === 'duration-asc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowDownIcon" cover stroke />
+                </span>
+              )}
+              {sortKey === 'duration-desc' && (
+                <span className={style.sortIcon}>
+                  <Icon icon="ArrowUpIcon" cover stroke />
+                </span>
+              )}
+            </div>
           </div>
         )}
 
@@ -113,7 +198,8 @@ const ListTracks = ({ variant, albumId, playlistId, entries }) => {
                   playingVariant: variant,
                   playingAlbumId: albumId,
                   playingPlaylistId: playlistId,
-                  playingTrackIndex: index,
+                  playingOrder: sortKey ? playingOrder : null,
+                  playingTrackIndex: sortKey ? playingOrder[index] : index,
                 });
               } else {
                 dispatch.playerModel.playerPlay();
@@ -227,6 +313,8 @@ const ListEntry = React.memo(
               <StarRating type="track" ratingKey={entry.trackId} rating={entry.userRating} editable />
             </div>
           )}
+
+          {/* {variant === 'playlists' && <div className={style.addedAt}>{addedAtToString(entry.addedAt)}</div>} */}
 
           <div className={style.duration}>{durationToStringShort(entry.duration)}</div>
         </div>