+++ src/js/components/index.js
@@ -1,5 +1,7 @@
 export { default as Button } from './Button/Button';
 export { default as ControlBar } from './ControlBar/ControlBar';
+export { default as FilterSelect } from './FilterSelect/FilterSelect';
+export { default as FilterWrap } from './FilterWrap/FilterWrap';
 export { default as Icon } from './Icon/Icon';
 export { default as ListCards } from './ListCards/ListCards';
 export { default as ListServers } from './ListServers/ListServers';
@@ -9,7 +11,6 @@ export { default as PageHome } from './PageHome/PageHome';
 export { default as Queue } from './Queue/Queue';
 export { default as RangeSlider } from './RangeSlider/RangeSlider';
 export { default as RightBar } from './RightBar/RightBar';
-export { default as Select } from './Select/Select';
 export { default as Settings } from './Settings/Settings';
 export { default as SideBar } from './SideBar/SideBar';
 export { default as StarRating } from './StarRating/StarRating';