+++ src/js/pages/AlbumList.js
@@ -5,7 +5,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { ListCards, Loading, Select, TitleHeading } from 'js/components';
+import { FilterSelect, FilterWrap, ListCards, Loading, TitleHeading } from 'js/components';
 import { sortList } from 'js/utils';
 import * as plex from 'js/services/plex';
 
@@ -19,11 +19,12 @@ const AlbumList = () => {
   const currentLibrary = useSelector(({ sessionModel }) => sessionModel.currentLibrary);
   const currentLibraryId = currentLibrary?.libraryId;
   const sortAlbums = useSelector(({ sessionModel }) => sessionModel.sortAlbums);
+  const orderAlbums = useSelector(({ sessionModel }) => sessionModel.orderAlbums);
 
   const allAlbums = useSelector(({ appModel }) => appModel.allAlbums)?.filter(
     (album) => album.libraryId === currentLibraryId
   );
-  const sortedAlbums = allAlbums ? sortList(allAlbums, sortAlbums) : null;
+  const sortedAlbums = allAlbums ? sortList(allAlbums, sortAlbums, orderAlbums) : null;
 
   useEffect(() => {
     plex.getAllAlbums();
@@ -35,24 +36,38 @@ const AlbumList = () => {
         title="Albums"
         subtitle={sortedAlbums ? sortedAlbums?.length + ' Album' + (sortedAlbums?.length !== 1 ? 's' : '') : null}
       />
-      <Select
-        value={sortAlbums}
-        options={[
-          { value: 'title', label: 'Alphabetical' },
-          { value: 'artist', label: 'Artist' },
-          { value: 'artist-asc-releaseDate-asc', label: 'Artist, by Newest Release' },
-          { value: 'artist-asc-releaseDate-desc', label: 'Artist, by Oldest Release' },
-          { value: 'userRating', label: 'Rating' },
-          { value: 'releaseDate', label: 'Release Date' },
-          { value: 'addedAt', label: 'Recently Added' },
-          { value: 'lastPlayed', label: 'Recently Played' },
-        ]}
-        setter={(sortAlbums) => {
-          dispatch.sessionModel.setSessionState({
-            sortAlbums,
-          });
-        }}
-      />
+      <FilterWrap>
+        <FilterSelect
+          value={sortAlbums}
+          options={[
+            { value: 'title', label: 'Alphabetical' },
+            { value: 'artist', label: 'Artist' },
+            { value: 'artist-asc-releaseDate-asc', label: 'Artist, by Newest Release' },
+            { value: 'artist-asc-releaseDate-desc', label: 'Artist, by Oldest Release' },
+            { value: 'userRating', label: 'Rating' },
+            { value: 'releaseDate', label: 'Release Date' },
+            { value: 'addedAt', label: 'Recently Added' },
+            { value: 'lastPlayed', label: 'Recently Played' },
+          ]}
+          setter={(sortAlbums) => {
+            dispatch.sessionModel.setSessionState({
+              sortAlbums,
+            });
+          }}
+        />
+        <FilterSelect
+          value={orderAlbums}
+          options={[
+            { value: 'asc', label: 'Ascending' },
+            { value: 'desc', label: 'Descending' },
+          ]}
+          setter={(orderAlbums) => {
+            dispatch.sessionModel.setSessionState({
+              orderAlbums,
+            });
+          }}
+        />
+      </FilterWrap>
       {!sortedAlbums && <Loading forceVisible inline />}
       {sortedAlbums && <ListCards variant="albums" entries={sortedAlbums} />}
     </>