+++ src/js/pages/ArtistCollectionList.js
@@ -5,7 +5,7 @@
 import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { ListCards, Loading, Select, TitleHeading } from 'js/components';
+import { FilterSelect, FilterWrap, ListCards, Loading, TitleHeading } from 'js/components';
 import { sortList } from 'js/utils';
 import * as plex from 'js/services/plex';
 
@@ -18,8 +18,11 @@ const ArtistCollectionList = () => {
 
   const allArtistCollections = useSelector(({ appModel }) => appModel.allArtistCollections);
   const sortArtistCollections = useSelector(({ sessionModel }) => sessionModel.sortArtistCollections);
+  const orderArtistCollections = useSelector(({ sessionModel }) => sessionModel.orderArtistCollections);
 
-  const sortedArtistCollections = allArtistCollections ? sortList(allArtistCollections, sortArtistCollections) : null;
+  const sortedArtistCollections = allArtistCollections
+    ? sortList(allArtistCollections, sortArtistCollections, orderArtistCollections)
+    : null;
 
   useEffect(() => {
     plex.getAllCollections();
@@ -37,19 +40,33 @@ const ArtistCollectionList = () => {
           )
         }
       />
-      <Select
-        value={sortArtistCollections}
-        options={[
-          { value: 'title', label: 'Alphabetical' },
-          { value: 'userRating', label: 'Rating' },
-          { value: 'addedAt', label: 'Recently Added' },
-        ]}
-        setter={(sortArtistCollections) => {
-          dispatch.sessionModel.setSessionState({
-            sortArtistCollections,
-          });
-        }}
-      />
+      <FilterWrap>
+        <FilterSelect
+          value={sortArtistCollections}
+          options={[
+            { value: 'title', label: 'Alphabetical' },
+            { value: 'userRating', label: 'Rating' },
+            { value: 'addedAt', label: 'Recently Added' },
+          ]}
+          setter={(sortArtistCollections) => {
+            dispatch.sessionModel.setSessionState({
+              sortArtistCollections,
+            });
+          }}
+        />
+        <FilterSelect
+          value={orderArtistCollections}
+          options={[
+            { value: 'asc', label: 'Ascending' },
+            { value: 'desc', label: 'Descending' },
+          ]}
+          setter={(orderArtistCollections) => {
+            dispatch.sessionModel.setSessionState({
+              orderArtistCollections,
+            });
+          }}
+        />
+      </FilterWrap>
       {!sortedArtistCollections && <Loading forceVisible inline />}
       {sortedArtistCollections && <ListCards variant="collections" entries={sortedArtistCollections} />}
     </>