+++ src/js/utils/sortList.js
@@ -1,20 +1,25 @@
-const sortList = (entry, options) => {
+const sortList = (entries, options, direction = 'asc') => {
   const optionsArray = options.split('-');
 
   const primarySortKey = optionsArray[0];
-  const primaryDirection = optionsArray[1] || 'asc';
+  let primaryDirection = optionsArray[1] || 'asc';
 
   const secondarySortKey = optionsArray[2] || 'title';
-  const secondaryDirection = optionsArray[3] || 'asc';
+  let secondaryDirection = optionsArray[3] || 'asc';
 
-  return doSorting(entry, primarySortKey, secondarySortKey, primaryDirection, secondaryDirection);
+  if (direction === 'desc') {
+    primaryDirection = primaryDirection === 'asc' ? 'desc' : 'asc';
+    secondaryDirection = secondaryDirection === 'asc' ? 'desc' : 'asc';
+  }
+
+  return doSorting(entries, primarySortKey, primaryDirection, secondarySortKey, secondaryDirection);
 };
 
 const doSorting = (
-  entry,
+  entries,
   primarySortKey,
-  secondarySortKey = 'title',
   primaryDirection = 'asc',
+  secondarySortKey = 'title',
   secondaryDirection = 'asc'
 ) => {
   primarySortKey = sortFunctions[primarySortKey] ? primarySortKey : 'title';
@@ -23,7 +28,7 @@ const doSorting = (
   const primaryDirectionFactor = primaryDirection === 'asc' ? 1 : -1;
   const secondaryDirectionFactor = secondaryDirection === 'asc' ? 1 : -1;
 
-  return entry.sort((a, b) => {
+  return [...entries].sort((a, b) => {
     const primaryComparison = primaryDirectionFactor * sortFunctions[primarySortKey](a, b);
     if (primaryComparison === 0 && secondarySortKey) {
       return secondaryDirectionFactor * sortFunctions[secondarySortKey](a, b);
@@ -33,6 +38,7 @@ const doSorting = (
 };
 
 const sortFunctions = {
+  // Strings
   title: (a, b) => {
     const nameA = a.title.toUpperCase();
     const nameB = b.title.toUpperCase();
@@ -44,11 +50,19 @@ const sortFunctions = {
     }
     return nameA.localeCompare(nameB);
   },
+  album: (a, b) => a.album?.localeCompare(b.album),
   artist: (a, b) => a.artist?.localeCompare(b.artist),
+
+  // Numbers
+  duration: (a, b) => a.duration - b.duration,
+  sortOrder: (a, b) => (parseInt(a.sortOrder) || 0) - (parseInt(b.sortOrder) || 0),
+  trackNumber: (a, b) => (parseInt(a.trackNumber) || 0) - (parseInt(b.trackNumber) || 0),
   userRating: (a, b) => (parseInt(b.userRating) || 0) - (parseInt(a.userRating) || 0),
-  releaseDate: (a, b) => new Date(b.releaseDate || '1970-01-01') - new Date(a.releaseDate || '1970-01-01'),
+
+  // Dates
   addedAt: (a, b) => new Date(b.addedAt || '1970-01-01') - new Date(a.addedAt || '1970-01-01'),
   lastPlayed: (a, b) => new Date(b.lastPlayed || '1970-01-01') - new Date(a.lastPlayed || '1970-01-01'),
+  releaseDate: (a, b) => new Date(b.releaseDate || '1970-01-01') - new Date(a.releaseDate || '1970-01-01'),
 };
 
 export default sortList;